Add comments to DagData trait
